backend/requirements.txt
fastapi[standard]==0.133.0
pwdlib[argon2]==0.3.0
sqlmodel==0.0.37
pyjwt==2.11.0
pymysql==1.1.2
cryptography==46.0.5